fix: 클래스 분리
클래스 분리
